pyproject.toml
[project]
name = "projeto-01-ingestao"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.42.84",
    "dlt[parquet,s3]>=1.24.0",
    "iterators>=0.2.0",
    "minio>=7.2.20",
    "pandas>=3.0.2",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
]
